Remove travis tests in favor of github actions
